% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istesting-package.R
\docType{package}
\name{istesting-package}
\alias{istesting}
\alias{istesting-package}
\title{istesting: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jennybc/istesting}
  \item Report bugs at \url{https://github.com/jennybc/istesting/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

}
\keyword{internal}
